+++ src/mjolnir/pbfgraphparser.cc
@@ -117,6 +117,10 @@ public:
     bool is_highway_junction =
         ((highway_junction != results->end()) && (highway_junction->second == "motorway_junction"));
 
+    const auto& junction_name = results->find("junction");
+    bool has_junction_name =
+        ((junction_name != results->end()) && (junction_name->second == "named"));
+
     // Create a new node and set its attributes
     OSMNode n;
     n.set_id(osmid);
@@ -147,7 +151,7 @@ public:
           n.set_ref_index(osmdata_.node_names.index(tag.second));
           ++osmdata_.node_ref_count;
         }
-      } else if (is_highway_junction && (tag.first == "name")) {
+      } else if ((is_highway_junction || has_junction_name) && (tag.first == "name")) {
         bool hasTag = (tag.second.length() ? true : false);
         if (hasTag) {
           // Add the name to the unique node names list and store its index in the OSM node
@@ -188,6 +192,8 @@ public:
         }
       } else if (tag.first == "access_mask") {
         n.set_access(std::stoi(tag.second));
+      } else if (has_junction_name) {
+        n.set_named_intersection(true);
       }
 
       /* TODO: payment type.
@@ -364,7 +370,6 @@ public:
         ((highway_junction != results.end()) && (highway_junction->second == "motorway_junction"));
 
     for (const auto& tag : results) {
-
       if (tag.first == "road_class") {
         RoadClass roadclass = (RoadClass)std::stoi(tag.second);
         switch (roadclass) {