+++ src/mjolnir/transitbuilder.cc
@@ -100,7 +100,7 @@ void ConnectToGraph(GraphTileBuilder& tilebuilder_local,
   // present in this tile set a value > number of directed edges
   uint32_t signidx = 0;
   uint32_t nextsignidx = (tilebuilder_local.header()->signcount() > 0)
-                             ? tilebuilder_local.sign(0).edgeindex()
+                             ? tilebuilder_local.sign(0).index()
                              : currentedges.size() + 1;
   uint32_t signcount = tilebuilder_local.header()->signcount();
 
@@ -127,14 +127,14 @@ void ConnectToGraph(GraphTileBuilder& tilebuilder_local,
       // Update any signs that use this idx - increment their index by the
       // number of added edges
       while (idx == nextsignidx && signidx < signcount) {
-        if (!currentedges[idx].exitsign()) {
+        if (!currentedges[idx].sign()) {
           LOG_ERROR("Signs for this index but directededge says no sign");
         }
-        tilebuilder_local.sign_builder(signidx).set_edgeindex(idx + added_edges);
+        tilebuilder_local.sign_builder(signidx).set_index(idx + added_edges);
 
         // Increment to the next sign and update nextsignidx
         signidx++;
-        nextsignidx = (signidx >= signcount) ? 0 : tilebuilder_local.sign(signidx).edgeindex();
+        nextsignidx = (signidx >= signcount) ? 0 : tilebuilder_local.sign(signidx).index();
       }
 
       // Add any restrictions that use this idx - increment their index by the
@@ -234,9 +234,8 @@ void ConnectToGraph(GraphTileBuilder& tilebuilder_local,
   // Get the directed edge index of the first sign. If no signs are
   // present in this tile set a value > number of directed edges
   signidx = 0;
-  nextsignidx = (tilebuilder_transit.header()->signcount() > 0)
-                    ? tilebuilder_transit.sign(0).edgeindex()
-                    : currentedges.size() + 1;
+  nextsignidx = (tilebuilder_transit.header()->signcount() > 0) ? tilebuilder_transit.sign(0).index()
+                                                                : currentedges.size() + 1;
   signcount = tilebuilder_transit.header()->signcount();
 
   // Get the directed edge index of the first access restriction.