+++ test/signinfo.cc
@@ -15,107 +15,171 @@ namespace {
 
 void ExitToTest() {
   OSMNode node{1234};
+  OSMNode exit_node{1234};
+
   OSMWay way{};
   OSMData osmdata{};
   bool fork = false;
   bool forward = true;
 
+  exit_node.set_ref_index(osmdata.node_names.index("5"));
+  exit_node.set_name_index(osmdata.node_names.index("PATP West Exit"));
+
+  std::vector<SignInfo> signs;
+  bool has_guide =
+      GraphBuilder::CreateSignInfoList(exit_node, way, osmdata, signs, fork, forward, true, false);
+
+  if (has_guide)
+    throw std::runtime_error("Exits should not be Guides");
+
+  if (signs.size() == 2) {
+    if (signs[0].type() != Sign::Type::kExitNumber)
+      throw std::runtime_error("Exit 5 is not an exit Number");
+    if (signs[0].text() != "5")
+      throw std::runtime_error("Exitsign text is bad for Exit 5.");
+    if (signs[1].type() != Sign::Type::kExitName)
+      throw std::runtime_error("PATP West Exit is not an exit Name");
+    if (signs[1].text() != "PATP West Exit")
+      throw std::runtime_error("Exitsign text is bad for PATP West Exit.");
+
+  } else
+    throw std::runtime_error("Exit 5 failed to parse.  " + std::to_string(signs.size()));
+
   node.set_exit_to_index(osmdata.node_names.index("US 11;To I 81;Carlisle;Harrisburg"));
+  signs.clear();
+  has_guide = GraphBuilder::CreateSignInfoList(node, way, osmdata, signs, fork, forward, true, false);
 
-  std::vector<SignInfo> exitsigns;
-  exitsigns = GraphBuilder::CreateExitSignInfoList(node, way, osmdata, fork, forward);
+  if (has_guide)
+    throw std::runtime_error("Exits should not be Guides");
 
-  if (exitsigns.size() == 4) {
-    for (auto& exitsign : exitsigns) {
+  if (signs.size() == 4) {
+    for (auto& exitsign : signs) {
       if (exitsign.type() != Sign::Type::kExitToward)
         throw std::runtime_error("US 11;To I 81;Carlisle;Harrisburg types are not all Toward");
     }
 
-    if (exitsigns[0].text() != "US 11" && exitsigns[1].text() != "I 81" &&
-        exitsigns[2].text() != "Carlisle" && exitsigns[3].text() != "Harrisburg")
+    if (signs[0].text() != "US 11" && signs[1].text() != "I 81" && signs[2].text() != "Carlisle" &&
+        signs[3].text() != "Harrisburg")
       throw std::runtime_error("Exitsign text is bad for US 11;To I 81;Carlisle;Harrisburg.");
   } else
     throw std::runtime_error("US 11/To I 81/Carlisle/Harrisburg failed to be parsed.  " +
-                             std::to_string(exitsigns.size()));
+                             std::to_string(signs.size()));
 
-  exitsigns.clear();
+  signs.clear();
   node.set_exit_to_index(osmdata.node_names.index("US 11;Toward I 81;Carlisle;Harrisburg"));
 
-  exitsigns = GraphBuilder::CreateExitSignInfoList(node, way, osmdata, fork, forward);
+  has_guide = GraphBuilder::CreateSignInfoList(node, way, osmdata, signs, fork, forward, true, false);
 
-  if (exitsigns.size() == 4) {
-    for (auto& exitsign : exitsigns) {
+  if (has_guide)
+    throw std::runtime_error("Exits should not be Guides");
+
+  if (signs.size() == 4) {
+    for (auto& exitsign : signs) {
       if (exitsign.type() != Sign::Type::kExitToward)
         throw std::runtime_error("US 11;Toward I 81;Carlisle;Harrisburg types are not all Toward");
     }
-    if (exitsigns[0].text() != "US 11" && exitsigns[1].text() != "I 81" &&
-        exitsigns[2].text() != "Carlisle" && exitsigns[3].text() != "Harrisburg")
+    if (signs[0].text() != "US 11" && signs[1].text() != "I 81" && signs[2].text() != "Carlisle" &&
+        signs[3].text() != "Harrisburg")
       throw std::runtime_error("Exitsign text is bad for US 11;To I 81;Carlisle;Harrisburg.");
   } else
     throw std::runtime_error("US 11;Toward I 81;Carlisle;Harrisburg failed to be parsed.");
 
-  exitsigns.clear();
+  signs.clear();
   node.set_exit_to_index(osmdata.node_names.index("I 95 To I 695"));
 
-  exitsigns = GraphBuilder::CreateExitSignInfoList(node, way, osmdata, fork, forward);
+  has_guide = GraphBuilder::CreateSignInfoList(node, way, osmdata, signs, fork, forward, true, false);
 
-  if (exitsigns.size() == 2) {
-    if (exitsigns[0].type() != Sign::Type::kExitBranch)
+  if (has_guide)
+    throw std::runtime_error("Exits should not be Guides");
+
+  if (signs.size() == 2) {
+    if (signs[0].type() != Sign::Type::kExitBranch)
       throw std::runtime_error("I 95 should be a branch.");
 
-    if (exitsigns[1].type() != Sign::Type::kExitToward)
+    if (signs[1].type() != Sign::Type::kExitToward)
       throw std::runtime_error("I 695 should be a toward.");
 
-    if (exitsigns[0].text() != "I 95" && exitsigns[1].text() != "I 695")
+    if (signs[0].text() != "I 95" && signs[1].text() != "I 695")
       throw std::runtime_error("Exitsign text is bad for I 95 To I 695");
   } else
     throw std::runtime_error("I 95 To I 695 failed to be parsed.");
 
-  exitsigns.clear();
+  signs.clear();
   node.set_exit_to_index(osmdata.node_names.index("I 495 Toward I 270"));
 
-  exitsigns = GraphBuilder::CreateExitSignInfoList(node, way, osmdata, fork, forward);
+  has_guide = GraphBuilder::CreateSignInfoList(node, way, osmdata, signs, fork, forward, true, false);
 
-  if (exitsigns.size() == 2) {
-    if (exitsigns[0].type() != Sign::Type::kExitBranch)
+  if (has_guide)
+    throw std::runtime_error("Exits should not be Guides");
+
+  if (signs.size() == 2) {
+    if (signs[0].type() != Sign::Type::kExitBranch)
       throw std::runtime_error("I 495 should be a branch.");
 
-    if (exitsigns[1].type() != Sign::Type::kExitToward)
+    if (signs[1].type() != Sign::Type::kExitToward)
       throw std::runtime_error("I 270 should be a toward.");
 
-    if (exitsigns[0].text() != "I 495" && exitsigns[1].text() != "I 270")
+    if (signs[0].text() != "I 495" && signs[1].text() != "I 270")
       throw std::runtime_error("Exitsign text is bad for I 495 Toward I 270");
   } else
     throw std::runtime_error("I 495 Toward I 270 failed to be parsed.");
 
-  exitsigns.clear();
+  signs.clear();
   node.set_exit_to_index(
       osmdata.node_names.index("I 495 Toward I 270 To I 95")); // default to toward.  Punt on parsing.
 
-  exitsigns = GraphBuilder::CreateExitSignInfoList(node, way, osmdata, fork, forward);
+  has_guide = GraphBuilder::CreateSignInfoList(node, way, osmdata, signs, fork, forward, true, false);
 
-  if (exitsigns.size() == 1) {
-    if (exitsigns[0].type() != Sign::Type::kExitToward)
+  if (has_guide)
+    throw std::runtime_error("Exits should not be Guides");
+
+  if (signs.size() == 1) {
+    if (signs[0].type() != Sign::Type::kExitToward)
       throw std::runtime_error("I 495 Toward I 270 To I 95 should be a toward.");
 
-    if (exitsigns[0].text() != "I 495 Toward I 270 To I 95")
+    if (signs[0].text() != "I 495 Toward I 270 To I 95")
       throw std::runtime_error("Exitsign text is bad for I 495 Toward I 270 To I 95");
   } else
     throw std::runtime_error("I 495 Toward I 270 To I 95 failed to be parsed.");
 
   // Add a ref branch sign
-  exitsigns.clear();
+  signs.clear();
   auto index = osmdata.name_offset_map.index("I 495 North");
   way.set_destination_ref_index(index);
-  exitsigns = GraphBuilder::CreateExitSignInfoList(node, way, osmdata, fork, forward);
-  if (exitsigns.size() == 1) {
-    if (exitsigns[0].type() != Sign::Type::kExitBranch)
+  has_guide = GraphBuilder::CreateSignInfoList(node, way, osmdata, signs, fork, forward, true, false);
+
+  if (has_guide)
+    throw std::runtime_error("Exits should not be Guides");
+
+  if (signs.size() == 1) {
+    if (signs[0].type() != Sign::Type::kExitBranch)
       throw std::runtime_error("I 495 North should be a branch.");
 
-    if (!exitsigns[0].is_route_num())
+    if (!signs[0].is_route_num())
       throw std::runtime_error("I 495 North should be flagged as a route num");
 
-    if (exitsigns[0].text() != "I 495 North")
+    if (signs[0].text() != "I 495 North")
+      throw std::runtime_error("Exitsign text is bad for I 495 North destination ref");
+  } else {
+    throw std::runtime_error("destination ref I 495 North failed to create exist sign.");
+  }
+
+  signs.clear();
+  index = osmdata.name_offset_map.index("I 495 North");
+  way.set_destination_ref_index(index);
+  has_guide = GraphBuilder::CreateSignInfoList(node, way, osmdata, signs, fork, forward, false, true);
+
+  if (!has_guide)
+    throw std::runtime_error("Guides should not be Exits");
+
+  if (signs.size() == 1) {
+    if (signs[0].type() != Sign::Type::kGuideBranch)
+      throw std::runtime_error("I 495 North should be a branch.");
+
+    if (!signs[0].is_route_num())
+      throw std::runtime_error("I 495 North should be flagged as a route num");
+
+    if (signs[0].text() != "I 495 North")
       throw std::runtime_error("Exitsign text is bad for I 495 North destination ref");
   } else {
     throw std::runtime_error("destination ref I 495 North failed to create exist sign.");
@@ -123,21 +187,73 @@ void ExitToTest() {
 
   // Add a ref toward sign
   OSMWay way2{};
-  exitsigns.clear();
+  signs.clear();
   auto index2 = osmdata.name_offset_map.index("I 695 North");
   way2.set_destination_ref_to_index(index2);
-  exitsigns = GraphBuilder::CreateExitSignInfoList(node, way2, osmdata, fork, forward);
-  if (exitsigns.size() == 1) {
-    if (exitsigns[0].type() != Sign::Type::kExitToward)
+  has_guide =
+      GraphBuilder::CreateSignInfoList(node, way2, osmdata, signs, fork, forward, true, false);
+
+  if (has_guide)
+    throw std::runtime_error("Exits should not be Guides");
+
+  if (signs.size() == 1) {
+    if (signs[0].type() != Sign::Type::kExitToward)
       throw std::runtime_error("I 695 North should be a toward.");
 
-    if (!exitsigns[0].is_route_num())
+    if (!signs[0].is_route_num())
       throw std::runtime_error("I 695 North should be flagged as a route num");
 
-    if (exitsigns[0].text() != "I 695 North")
+    if (signs[0].text() != "I 695 North")
       throw std::runtime_error("Exitsign text is bad for I 695 North destination to ref");
   } else {
     throw std::runtime_error("destination ref I 695 North failed to create exist sign.");
+  }
+
+  // Add a ref toward guide sign
+  OSMWay way3{};
+  signs.clear();
+  auto index3 = osmdata.name_offset_map.index("I 695 North");
+  way3.set_destination_ref_to_index(index3);
+  has_guide =
+      GraphBuilder::CreateSignInfoList(node, way2, osmdata, signs, fork, forward, false, true);
+
+  if (!has_guide)
+    throw std::runtime_error("Guides should not be Exits");
+
+  if (signs.size() == 1) {
+    if (signs[0].type() != Sign::Type::kGuideToward)
+      throw std::runtime_error("I 695 North should be a toward.");
+
+    if (!signs[0].is_route_num())
+      throw std::runtime_error("I 695 North should be flagged as a route num");
+
+    if (signs[0].text() != "I 695 North")
+      throw std::runtime_error("Exitsign text is bad for I 695 North destination to ref");
+  } else {
+    throw std::runtime_error("destination ref I 695 North failed to create exist sign.");
+  }
+
+  // Add a ref toward guide sign and we should not add a exit number or exit name.  note: using
+  // exit_node
+  signs.clear();
+  has_guide =
+      GraphBuilder::CreateSignInfoList(exit_node, way2, osmdata, signs, fork, forward, false, true);
+
+  if (!has_guide)
+    throw std::runtime_error("Guides should not be Exits");
+
+  if (signs.size() == 1) {
+    if (signs[0].type() != Sign::Type::kGuideToward)
+      throw std::runtime_error("I 695 North should be a toward. No exit 5 should exist.");
+
+    if (!signs[0].is_route_num())
+      throw std::runtime_error("I 695 North should be flagged as a route num");
+
+    if (signs[0].text() != "I 695 North")
+      throw std::runtime_error("Exitsign text is bad for I 695 North destination to ref");
+  } else {
+    throw std::runtime_error(
+        "destination ref I 695 North failed to create exist sign.  No exit 5 should exist.");
   }
 }