+++ valhalla/mjolnir/graphbuilder.h
@@ -45,11 +45,14 @@ public:
 
   static std::string GetRef(const std::string& way_ref, const std::string& relation_ref);
 
-  static std::vector<baldr::SignInfo> CreateExitSignInfoList(const OSMNode& node,
-                                                             const OSMWay& way,
-                                                             const OSMData& osmdata,
-                                                             bool fork,
-                                                             bool forward);
+  static bool CreateSignInfoList(const OSMNode& node,
+                                 const OSMWay& way,
+                                 const OSMData& osmdata,
+                                 std::vector<baldr::SignInfo>& exits,
+                                 bool fork,
+                                 bool forward,
+                                 bool ramp,
+                                 bool tc);
 };
 
 } // namespace mjolnir