+++ valhalla/odin/maneuversbuilder.h
@@ -60,7 +60,7 @@ protected:
                                                  std::list<Maneuver>::iterator curr_man,
                                                  std::list<Maneuver>::iterator next_man);
 
-  void CountAndSortExitSigns(std::list<Maneuver>& maneuvers);
+  void CountAndSortSigns(std::list<Maneuver>& maneuvers);
 
   void ConfirmManeuverTypeAssignment(std::list<Maneuver>& maneuvers);